Fix time bug for the shorter February month
Ruby's subtraction of dates took into account the shorter February
month.  When you compare this result with `1.month` (ie. 30 days in
Rubyland) you get an incorrect comparison and a failing test.

This creates a small exception for February, removes duplicate code and
adds a February specific test in.
