+++ app/controllers/administrator_reports_controller.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class AdministratorReportsController < ApplicationController
+  include Steps
+
   before_action { check_access! 'admin:report' }
   before_action :set_range_params,
                 except: %i[subscriber_registrations_report index]
@@ -48,32 +50,5 @@ class AdministratorReportsController < ApplicationController
 
     report = ServiceProviderUtilizationReport.new(start, finish)
     @data = JSON.generate(report.generate)
-  end
-
-  private
-
-  def set_range_params
-    @start = params[:start]
-    @end = params[:end]
-  end
-
-  def start
-    return nil if params[:start].blank?
-    Time.zone.parse(params[:start]).beginning_of_day
-  end
-
-  def finish
-    return nil if params[:end].blank?
-    Time.zone.parse(params[:end]).tomorrow.beginning_of_day
-  end
-
-  def scaled_steps
-    range = finish - start
-
-    return 24 if range >= 1.year
-    return 12 if range >= 6.months
-    return 6 if range >= 3.months
-    return 2 if range >= 1.month
-    1
   end
 end
