+++ app/controllers/subscriber_reports.rb
@@ -1,6 +1,8 @@
 # frozen_string_literal: true
 
 class SubscriberReports < ApplicationController
+  include Steps
+
   before_action { permitted_objects(model_object) }
   before_action :requested_entity
   before_action :set_range_params
@@ -43,30 +45,5 @@ class SubscriberReports < ApplicationController
     @entities = active_objects.select do |sp|
       subject.permits? permission_string(sp)
     end
-  end
-
-  def set_range_params
-    @start = params[:start]
-    @end = params[:end]
-  end
-
-  def start
-    return nil if params[:start].blank?
-    Time.zone.parse(params[:start]).beginning_of_day
-  end
-
-  def finish
-    return nil if params[:end].blank?
-    Time.zone.parse(params[:end]).tomorrow.beginning_of_day
-  end
-
-  def scaled_steps
-    range = finish - start
-
-    return 24 if range >= 1.year
-    return 12 if range >= 6.months
-    return 6 if range >= 3.months
-    return 2 if range >= 1.month
-    1
   end
 end
