+++ spec/support/scaling_steps.rb
@@ -15,6 +15,20 @@ RSpec.shared_examples 'report with scalable steps' do
     { start: Time.now.utc - 1.month, end: Time.now.utc }
   end
 
+  let(:range_1_month_february_exception) do
+    {
+      start: DateTime.parse('2017-02-09 00:00:00').utc,
+      end: DateTime.parse('2017-03-09 00:00:00').utc
+    }
+  end
+
+  let(:range_2_months_february_exception) do
+    {
+      start: DateTime.parse('2017-01-09 00:00:00').utc,
+      end: DateTime.parse('2017-03-09 00:00:00').utc
+    }
+  end
+
   let(:range_2_months) do
     { start: Time.now.utc - 2.months, end: Time.now.utc }
   end
@@ -63,15 +77,32 @@ RSpec.shared_examples 'report with scalable steps' do
     end
   end
 
-  it 'Steps should be 2 hours within 1 to 3 months' do
-    [range_1_month, range_2_months].each do |rng|
-      post path, params: params.merge(rng)
+  context 'Steps should be 2 hours within 1 to 3 months' do
+    it 'normal' do
+      [range_1_month, range_2_months].each do |rng|
+        post path, params: params.merge(rng)
 
-      data = JSON.parse(assigns[:data], symbolize_names: true)
-      sessions = data[:data][:sessions]
+        data = JSON.parse(assigns[:data], symbolize_names: true)
+        sessions = data[:data][:sessions]
 
-      expect(sessions[1]).to eq([7200, 0.0])
-      expect(sessions[2]).to eq([7200 * 2, 0.0])
+        expect(sessions[1]).to eq([7200, 0.0])
+        expect(sessions[2]).to eq([7200 * 2, 0.0])
+      end
+    end
+
+    it 'February exception' do
+      [
+        range_1_month_february_exception,
+        range_2_months_february_exception
+      ].each do |rng|
+        post path, params.merge(rng)
+
+        data = JSON.parse(assigns[:data], symbolize_names: true)
+        sessions = data[:data][:sessions]
+
+        expect(sessions[1]).to eq([7200, 0.0])
+        expect(sessions[2]).to eq([7200 * 2, 0.0])
+      end
     end
   end
